\name{createHeader}
\alias{createHeader}
\title{Create the header of the page}
\usage{
  createHeader(all_options)
}
\arguments{
  \item{all_options}{The full set of options. Will
  generally be .all_options}
}
\description{
  This function takes the full set of blog options, and
  uses that to create the header that will be displayed on
  the main index page and each blog post page
}

