\name{createMain}
\alias{createMain}
\title{Create the index page of recent blog posts}
\usage{
  createMain(all_options)
}
\arguments{
  \item{all_options}{The full set of options}
}
\description{
  Takes all_options and creates the index page of recent
  posts. The number of posts, and many other options should
  be set in the file _blog_options.R. This file will be
  sourced, and the options used in subsequent creation of
  the main index and individual posts.
}

