\name{createPost}
\alias{createPost}
\title{Create an individual post}
\usage{
  createPost(filename, all_options)
}
\arguments{
  \item{all_options}{the options for the blog generally}

  \item{filename}{the MD file to be processed to generate
  the individual post}
}
\description{
  Takes a supplied file \code{filename} and generates an
  html page based on a given template
}

