\name{do.blogR}
\alias{do.blogR}
\title{Process and post the latest blog entry.}
\description{Process and post the latest blog entry.}
\details{Function to process and post a blog entry using all the global enties.}
\value{character The url for the posted entry.}
\seealso{\code{\link{process.blog}}, \code{\link{post.blog}}}
\author{Shane Conway \email{shane.conway@gmail.com}}
\keyword{character}
