% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_posts.R
\name{ensure_posts}
\alias{ensure_posts}
\title{Ensure that the \code{posts} directory exists}
\usage{
ensure_posts(blogdir = ".")
}
\arguments{
\item{blogdir}{path to the base directory of the blog}
}
\value{
The directory given by \code{paste(blogdir, "posts", sep = "/")} is 
created if it does not already exist.
}
\description{
\code{ensure_posts} ensures the existence of the \code{posts} 
directory
}
\examples{
ensure_posts()
}
\seealso{
\code{\link{new_post}}
}

