% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_shell.R
\name{ensure_shell}
\alias{ensure_shell}
\title{Ensure that the \code{shell} Rmd document exists}
\usage{
ensure_shell(blogdir = ".")
}
\arguments{
\item{blogdir}{path to the base directory of the blog}
}
\value{
The file given by \code{paste(blogdir, "shell.Rmd", sep = "/")} is 
created if it does not already exist.
}
\description{
\code{ensure_shell} ensures the existence of the \code{shell} 
Rmd document
}
\examples{
ensure_shell()
}
\seealso{
\code{\link{ensure_posts}}
}

