% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_info.R
\name{post_info}
\alias{post_info}
\title{Fetch the information about a post}
\usage{
post_info(post)
}
\arguments{
\item{post}{path to the post}
}
\value{
A named list containing information about \code{post}
}
\description{
\code{post_info} fetches the information about a post and 
returns it in a named list
}
\examples{
new_post("Hello, world!", "hello-world")
tag_post(post_dir(Sys.Date(), "hello-world"), c("greetings", "salutations"))
post_info(post_dir(Sys.Date(), "hello-world"))
}
\seealso{
\code{\link{new_post}, \link{tag_post}}
}

