% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_search.R
\name{post_search}
\alias{post_search}
\title{Generate a search table for blog posts}
\usage{
post_search(blogdir = ".", linkdir = "/blog")
}
\arguments{
\item{blogdir}{path to the base directory of the blog}

\item{linkdir}{relative path to the blog directory to use in post hyperlinks}
}
\value{
A \code{\link{datatable}} containing post information
}
\description{
\code{post_search} creates a searchable table of post information
}
\examples{
lapply(seq.Date(Sys.Date() - 4, Sys.Date(), 1),
       new_post,
       title = "Lorem Ipsum", 
       name = "lorem-ipsum")
tag_post(get_post_dirs()[1], c("hello", "world"))
post_search()
}
\seealso{
\code{\link{post_info}}
}

