% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_post.R
\name{remove_post}
\alias{remove_post}
\title{Remove a post}
\usage{
remove_post(post)
}
\arguments{
\item{post}{path to the post}
}
\value{
The directory \code{post} is removed if it exists.
}
\description{
\code{remove_post} removes a post
}
\examples{
new_post("Hello, world!", "hello-world")
remove_post(post_dir(Sys.Date(), "hello-world"))
}
\seealso{
\code{\link{new_post}, \link{tag_post}, \link{render_post}}
}

