% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_post.R
\name{tag_post}
\alias{tag_post}
\title{Tag a post}
\usage{
tag_post(post, tags, append = TRUE)
}
\arguments{
\item{post}{path to the post}

\item{tags}{character vector of tags to apply}

\item{append}{logical value; add to existing tags (or replace)?}
}
\value{
The \code{tags} file in the directory \code{post} is created if it 
does not already exist, and populated with the items in \code{tags}.
}
\description{
\code{tag_post} tags a post
}
\examples{
new_post("Hello, world!", "hello-world")
tag_post(post_dir(Sys.Date(), "hello-world"), c("greetings", "salutations"))
}
\seealso{
\code{\link{new_post}, \link{render_post}, \link{remove_post}}
}

