\name{do.blogR}
\alias{do.blogR}
\title{Process and post the latest blog entry.}
\description{Process and post the latest blog entry.}
\details{Function to process and post a blog entry using all the global enties.}
\value{boolean vector}
\seealso{\code{\link{process.blog}}, \code{\link{post.blog}}}
\keyword{character}
\arguments{\item{string}{input character vector}
\item{pattern}{pattern to look for, as defined by a POSIX regular
expression.  See the ``Extended Regular Expressions'' section of 
\code{\link{regex}} for details.}}
