% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_post.R
\name{new_post}
\alias{new_post}
\title{Create a new post}
\usage{
new_post(title, name, date = Sys.Date(), blogdir = ".", ...)
}
\arguments{
\item{title}{the title of the post}

\item{name}{the name of the post}

\item{date}{the date of the post}

\item{blogdir}{path to the base directory of the blog}

\item{...}{ordered strings for substitution into \code{shell.Rmd}}
}
\value{
A new post is created in the \code{posts} directory using the 
\code{title}, \code{date} and \code{name} that were supplied.
}
\description{
\code{new_post} creates a new post
}
\examples{
new_post("Hello, world!", "hello-world")
}
\seealso{
\code{\link{tag_post}, \link{render_post}, \link{remove_post}}
}

