% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_post.R
\name{render_post}
\alias{render_post}
\title{Render a post to HTML}
\usage{
render_post(post)
}
\arguments{
\item{post}{path to the post}
}
\value{
The post HTML file is created.
}
\description{
\code{render_post} renders a post to HTML
}
\examples{
new_post("Hello, world!", Sys.Date(), "hello-world")
render_post(post_dir(Sys.Date(), "hello-world"))
}
\seealso{
\code{\link{new_post}}
}

